Updates the documentation Added new features : - getConfig - setConfig - deleteConfig - getMembership - Updated the admin class since it was outdated with the new cluster system. - Updated tests
Fixed test config
